build fixed. Doubles in dto replaced with bigdecimal

+++ store-dto/src/main/java/hr/ivlahek/sample/store/client/product/CreateProductDto.java
@@ -2,7 +2,6 @@ package hr.ivlahek.sample.store.client.product;
 
 import hr.ivlahek.sample.store.client.docs.DocumentationConstants;
 import hr.ivlahek.sample.store.client.validation.ValidationMessages;
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,6 +9,7 @@ import lombok.NoArgsConstructor;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
+import java.math.BigDecimal;
 
 @NoArgsConstructor
 @Data
@@ -28,5 +28,5 @@ public class CreateProductDto {
     @NotNull(message = ValidationMessages.PRODUCT_PRICE_NULL)
     @Positive(message = ValidationMessages.PRODUCT_PRICE_NEGATIVE_OR_ZERO)
     @ApiModelProperty(value = DocumentationConstants.PRODUCT_PRICE, required = true)
-    private Double price;
+    private BigDecimal price;
 }
